Update
Star rating added
Dropdown list to select a ticket to write a review added
Snippets and extras tag/filter added

+++ LITReview/reviews/views.py
@@ -13,10 +13,14 @@ from . import models, forms
 @login_required
 def home(request):
     tickets = models.Ticket.objects.filter(
-        Q(author__in=request.user.get_connections()) | Q(author=request.user)
+        Q(
+            author__in=request.user.get_connections()
+            ) | Q(author=request.user)
     )
     reviews = models.Review.objects.filter(
-        Q(user__in=request.user.get_connections()) | Q(user=request.user)
+        Q(
+            user__in=request.user.get_connections()
+            ) | Q(user=request.user)
     )
 
     flux = sorted(
@@ -24,7 +28,7 @@ def home(request):
         key=lambda instance: instance.time_created,
         reverse=True
     )
-    paginator = Paginator(flux, 6)
+    paginator = Paginator(flux, 5)
     page = request.GET.get('page')
     page_obj = paginator.get_page(page)
 
@@ -42,7 +46,7 @@ def personal_posts(request):
         Q(user=request.user)
     )
     flux = sorted(
-        chain(tickets, reviews),
+        chain(tickets, reviews), 
         key=lambda instance: instance.time_created,
         reverse=True
     )
@@ -53,26 +57,30 @@ def personal_posts(request):
     context = {
         'page_obj':page_obj, 
     }
-    return render(request, 'reviews/personal_posts.html', context=context)
+    return render(
+        request,
+        'reviews/personal_posts.html',
+        context=context
+        )
 
 class CreateTicketView(LoginRequiredMixin, CreateView):
     model = models.Ticket
-    fields = ['title', 'description', 'image']
+    form_class = forms.TicketForm
     template_name = 'reviews/create_ticket.html'
-    succes_url = '/'
+    success_url = '/'
 
-    def get_queryset(self):
-        queryset = super().get_queryset()
-        return queryset.filter(user=self.request.user)
+    def form_valid(self, form):
+        instance = form.save(commit=False)
+        instance.author = self.request.user
+        return super().form_valid(form)
 
 class TicketDetailsView(LoginRequiredMixin, DetailView):
     model = models.Ticket
-    fields = ['title', 'description', 'image']
     template_name = 'reviews/ticket_details.html'
 
 class UpdateTicketView(LoginRequiredMixin, UpdateView):
     model = models.Ticket
-    fields = ['title', 'description', 'image']
+    form_class = forms.TicketForm
     template_name = 'reviews/edit_ticket.html'
     success_url = '/'
 
@@ -81,38 +89,38 @@ class DeleteTicketView(LoginRequiredMixin, DeleteView):
     template_name = 'reviews/delete_ticket.html'
     success_url = '/'
 
-""" class CreateReviewView(LoginRequiredMixin, CreateView):
-    model = models.Ticket
-    fields = ['rating', 'headline', 'body']
-    template_name = 'reviews/create_review.html'
-
-class ReviewDetailsView(LoginRequiredMixin, DetailView):
-    model = models.Review
-    fields = ['rating', 'headline', 'body']
-    template_name = 'reviews/review_details.html'
-
-class UpdateReviewView(LoginRequiredMixin, UpdateView):
-    model = models.Review
-    fields =['rating', 'headline', 'body']
-    template_name = 'reviews/edit_review.html'
-    success_url = '/'
-
-class DeleteReviewView(LoginRequiredMixin, DeleteView):
-    model = models.Review
-    template_name = 'reviews/delete_review.html'
-    success_url = '/' """
-
 @login_required
 def create_review(request, ticket_id):
     ticket = get_object_or_404(models.Ticket, id=ticket_id)
+    tickets_list = models.Ticket.objects.filter(
+        Q(has_review=False) & (
+            Q(author__in=request.user.get_connections()) |
+            Q(author=request.user)
+            )
+    )
     form = forms.ReviewForm()
     if request.method == 'POST':
         form = forms.ReviewForm(request.POST)
         if form.is_valid():
             review = form.save(commit=False)
-            review.save(ticket, request.user)
-            return render(request, 'reviews/review_details.html', {'review':review})
-    return render(request, 'reviews/create_review.html', context={'form': form, 'ticket': ticket})
+            review.user = request.user
+            review.ticket = ticket
+            review.save()
+            return render(
+                request,
+                'reviews/review_details.html',
+                {'review':review}
+                )
+    context = {
+        'form':form,
+        'ticket':ticket,
+        'tickets_list':tickets_list
+    }
+    return render(
+        request,
+        'reviews/create_review.html',
+        context=context,
+        )
 
 @login_required
 def review_details(request, review_id):
@@ -133,21 +141,53 @@ def edit_review(request, review_id):
                 edit_form.save()
                 return redirect('home')
         if 'delete_review' in request.POST:
-            delete_form = forms.DeleteReviewForm(request.POST)
-            if delete_form.is_valid():
-                review.ticket.has_review=False
-                review.delete()
-                return redirect('home')
+            return render(
+                request.GET,
+                'reviews/delete_review.html',
+                context={'review':review},
+            )
     context = {
         'edit_form': edit_form,
         'delete_form': delete_form,
     }
-    return render(request, 'reviews/edit_review.html', context=context)
+    return render(
+        request,
+        'reviews/edit_review.html',
+        context=context
+        )
+
+@login_required
+def delete_review(request, review_id):
+    review = get_object_or_404(models.Review, id=review_id)
+    if request.user != review.user:
+        return redirect('review_details', review.id)
+    delete_form = forms.DeleteReviewForm()
+    if request.method == 'POST' and 'delete_review' in request.POST:
+        delete_form = forms.DeleteReviewForm(request.POST)
+        if delete_form.is_valid():
+            review.ticket.has_review=False
+            review.delete()
+            return redirect('home')
+    context = {
+        'review': review,
+        'delete_form': delete_form,
+    }
+    return render(
+        request,
+        'reviews/delete_review.html',
+        context={'review':review}
+        )
 
 @login_required
 def create_ticket_and_review(request):
     ticket_form = forms.TicketForm()
     review_form = forms.ReviewForm()
+    tickets_list = models.Ticket.objects.filter(
+        Q(has_review=False) & (
+            Q(author__in=request.user.get_connections()) |
+            Q(author=request.user)
+            )
+    )
     if request.method == 'POST':
         ticket_form = forms.TicketForm(request.POST, request.FILES)
         review_form = forms.ReviewForm(request.POST)
@@ -156,13 +196,20 @@ def create_ticket_and_review(request):
             ticket.author = request.user
             ticket.save()
             review = review_form.save(commit=False)
-            review.save(ticket, request.user)
+            review.user = request.user
+            review.ticket = ticket
+            review.save()
             return redirect('home')
     context = {
         'ticket_form': ticket_form,
-        'review_form': review_form
+        'review_form': review_form,
+        'tickets_list':tickets_list,
     }
-    return render(request, 'reviews/create_ticket_and_review.html', context=context)
+    return render(
+        request,
+        'reviews/create_ticket_and_review.html',
+        context=context
+        )
 
 @login_required
 def follow_users(request):
@@ -186,4 +233,8 @@ def follow_users(request):
                 form.clean_followed_user()
                 form.save()
                 return redirect('follow_users')
-    return render(request, 'reviews/follow_users_form.html', context={'form': form, 'delete_form': delete_form})
+    return render(
+        request,
+        'reviews/follow_users_form.html',
+        context={'form': form, 'delete_form': delete_form}
+        )
